pyproject.toml
[project]
name = "base-fast"
version = "0.1.0"
requires-python = "==3.13"
dependencies = [
    "fastapi==0.116.0",
    "httpx==0.28.1",
    "psycopg[binary,pool]==3.2.9",
    "pydantic==2.11.7",
    "passlib[bcrypt]==1.7.4",
    "python-jose==3.3.0",
    "pytest==8.4.1",
    "uvicorn==0.35.0",
    "python-multipart==0.0.20",
    "pytest-xdist==3.8.0",
]
